update to deicaml

+++ database/migrations/2023_04_13_143058_add_hours_per_week_to_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('hours_per_week')->nullable();
+            $table->decimal('hours_per_week')->nullable();
         });
     }
